querydsl 등록상품리스트 querydsl

+++ src/main/java/com/green/winey_final/admin/model/ProductInsParam.java
@@ -17,16 +17,16 @@ public class ProductInsParam {
 
     private Long country; //원산지
 
-    private Long sweety; //당도
-    private Long acidity; //산도
-    private Long body; //바디
+    private int sweety; //당도
+    private int acidity; //산도
+    private int body; //바디
 
     private Long category; //레드(1) 화이트(2) 스파클링(3) 기타(4) //
 
     private List<Long> aroma; //향
 
-    private Long sale; //할인률
-    private Long salePrice; //할인가격
+    private int sale; //할인률
+    private int salePrice; //할인가격
     private String startSale; //할인 시작일
     private String endSale; //할인 종료일
